Enable predeclared golang linter
predeclared linter finds code that shadows one of Go's predeclared identifiers.

Renamed cmd/clusterawsadm/cmd/ami/copy package to cmd/clusterawsadm/cmd/ami/common package
as copy is a predeclared identifier in Go.
