+++ cmd/clusterawsadm/cmd/ami/list/list.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package copy
+package list
 
 import (
 	"fmt"
